+++ apps/web/package.json
@@ -17,13 +17,13 @@
   },
   "dependencies": {
     "@rallly/backend": "*",
+    "@rallly/database": "*",
+    "@rallly/emails": "*",
+    "@rallly/tailwind-config": "*",
     "@floating-ui/react-dom-interactions": "^0.13.3",
     "@headlessui/react": "^1.7.7",
     "@next/bundle-analyzer": "^12.3.4",
     "@radix-ui/react-popover": "^1.0.3",
-    "@rallly/database": "*",
-    "@rallly/emails": "*",
-    "@rallly/tailwind-config": "*",
     "@sentry/nextjs": "^7.33.0",
     "@svgr/webpack": "^6.5.1",
     "@tailwindcss/typography": "^0.5.9",
