+++ apps/web/src/components/poll/mutations.ts
@@ -1,4 +1,4 @@
-import { trpc } from "@rallly/backend/next/trpc";
+import { trpc } from "@rallly/backend/client/trpc";
 
 import { usePostHog } from "@/utils/posthog";
