+++ apps/web/src/components/poll/notifications-toggle.tsx
@@ -1,4 +1,4 @@
-import { trpc } from "@rallly/backend/next/trpc";
+import { trpc } from "@rallly/backend/client/trpc";
 import { useTranslation } from "next-i18next";
 import * as React from "react";
