+++ apps/web/src/components/poll/use-touch-beacon.ts
@@ -1,4 +1,4 @@
-import { trpc } from "@rallly/backend/next/trpc";
+import { trpc } from "@rallly/backend/client/trpc";
 import { useMount } from "react-use";
 
 /**
