+++ apps/web/src/middleware.ts
@@ -1,4 +1,4 @@
-import { sessionConfig } from "@rallly/backend/session/config";
+import { sessionConfig } from "@rallly/backend/server/session";
 import languageParser from "accept-language-parser";
 import { getIronSession } from "iron-session/edge";
 import { NextRequest, NextResponse } from "next/server";
