+++ apps/web/src/pages/_app.tsx
@@ -2,8 +2,8 @@ import "react-big-calendar/lib/css/react-big-calendar.css";
 import "tailwindcss/tailwind.css";
 import "~/style.css";
 
-import { trpc } from "@rallly/backend/next/trpc";
-import { UserSession } from "@rallly/backend/trpc";
+import { UserSession } from "@rallly/backend";
+import { trpc } from "@rallly/backend/client/trpc";
 import { inject } from "@vercel/analytics";
 import { NextPage } from "next";
 import { AppProps } from "next/app";
