+++ apps/web/src/pages/demo.tsx
@@ -1,5 +1,8 @@
-import { withAuthIfRequired, withSessionSsr } from "@rallly/backend/next";
-import { trpc } from "@rallly/backend/next/trpc";
+import { trpc } from "@rallly/backend/client/trpc";
+import {
+  withAuthIfRequired,
+  withSessionSsr,
+} from "@rallly/backend/server/session";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
