+++ apps/web/src/pages/login.tsx
@@ -1,4 +1,4 @@
-import { withSessionSsr } from "@rallly/backend/next";
+import { withSessionSsr } from "@rallly/backend/server/session";
 import { GetServerSideProps, NextPage } from "next";
 import Head from "next/head";
 import { useRouter } from "next/router";
