+++ apps/web/src/utils/trpc/hooks.ts
@@ -1,4 +1,4 @@
-import { trpc } from "@rallly/backend/next/trpc";
+import { trpc } from "@rallly/backend/client/trpc";
 import { useRouter } from "next/router";
 
 export const usePollByAdmin = () => {
