+++ packages/backend/index.ts
@@ -1,2 +1,1 @@
-export * from "./session";
-export * from "./trpc";
+export * from "./types";
