+++ packages/backend/server/session/index.ts
@@ -1,2 +1,2 @@
+export * from "./next";
 export * from "./session";
-export * from "./utils";
