+++ packages/backend/server/session/session.ts
@@ -1,6 +1,13 @@
-import { sealData, unsealData } from "iron-session";
+import { IronSessionOptions, sealData, unsealData } from "iron-session";
 
-import { sessionConfig } from "./config";
+export const sessionConfig: IronSessionOptions = {
+  password: process.env.SECRET_PASSWORD ?? "",
+  cookieName: "rallly-session",
+  cookieOptions: {
+    secure: process.env.NODE_ENV === "production",
+  },
+  ttl: 60 * 60 * 24 * 30, // 30 days
+};
 
 export const decryptToken = async <P extends Record<string, unknown>>(
   token: string,
