+++ packages/backend/server/trpc/routers/polls/demo.ts
@@ -1,7 +1,7 @@
 import { prisma, VoteType } from "@rallly/database";
 import dayjs from "dayjs";
 
-import { nanoid } from "../../../utils/nanoid";
+import { nanoid } from "../../../../utils/nanoid";
 import { possiblyPublicProcedure, router } from "../..";
 
 const participantData: Array<{ name: string; votes: VoteType[] }> = [
