+++ packages/backend/server/trpc/routers/polls/participants.ts
@@ -4,7 +4,8 @@ import { absoluteUrl } from "@rallly/utils";
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
-import { createToken, DisableNotificationsPayload } from "../../../session";
+import { DisableNotificationsPayload } from "../../../../types";
+import { createToken } from "../../../session";
 import { publicProcedure, router } from "../..";
 
 export const participants = router({
