--- packages/backend/session/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./token";
-export * from "./types";
