--- packages/backend/trpc/types.ts
@@ -1,13 +1,0 @@
-export type RegisteredUserSession = {
-  isGuest: false;
-  id: string;
-  name: string;
-  email: string;
-};
-
-export type GuestUserSession = {
-  isGuest: true;
-  id: string;
-};
-
-export type UserSession = GuestUserSession | RegisteredUserSession;
