+++ packages/backend/types.ts
@@ -22,3 +22,17 @@ export type DisableNotificationsPayload = {
   pollId: string;
   watcherId: number;
 };
+
+export type RegisteredUserSession = {
+  isGuest: false;
+  id: string;
+  name: string;
+  email: string;
+};
+
+export type GuestUserSession = {
+  isGuest: true;
+  id: string;
+};
+
+export type UserSession = GuestUserSession | RegisteredUserSession;
